refactor(ImageCroppingWidget): Split out behavior and state
